requirements.txt
annotated-types==0.5.0
asgiref==3.7.2
defusedxml==0.7.1
diff-match-patch==20230430

django-filter==23.2
django-import-export==3.2.0
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
Pillow==10.0.0
pydantic==2.0.3
pydantic_core==2.3.0
PyYAML==6.0.1
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.7.1
tzdata==2023.3
xlrd==2.0.1
xlwt==1.3.0
